CHROMIUM: dm-verity: double bio_put() on backing disk requests
(Note, this is a first pass. It seems to work, but I'll cross check it more
 thoroughly tomorrow!)

When reading from the backing drive, a request will be cloned and optionally
padded.  Cloning always happens, if a request is padded, it is cloned twice.

On error, we were automatically calling bio_put() on the bio in the IO context
and the one that had the error even when it wasn't double cloned.  This change
adds the test I use elsewhere to see if the base_bio is the same as the one in
the IO context.  If it is, then the request was single-cloned(), if not, we need
to have the extra bio_put().

TEST=
1. created a new recovery kernel with mod_image_for_recovery.sh --kernel_image_only
2. Ran make_developer_script_runner.sh --developer_script shwrapper.sh
3. dd'd the output to a usb stick
4. Booted in dev mode
5. chrooted to my local system's /dev/sda3
6. create a verity table and hash file in /dev/sda1 for /dev/sdb1
7. created a loop0 for the hashes
8. Pulled out the usb stick and put it back in so it'd be /dev/sdc
9. Setup the dm device with dmsetup create -r foo --table "table output"
10. dd if=/dev/dm-0 of=/dev/null
11. echo 3 > /proc/sys/vm/drop_caches
12. dd if=/dev/dm-0 of=/dev/null bs=513 skip=2  and pull the stick out right after hitting enter
13. Check dmesg | tail.  The Error occurred message is there, but it successfully recovers.!
BUG=chromium-os:8499

Change-Id: I6b5a89a7de75028724e2a177b871bf02b766bc4b
Signed-off-by: Will Drewry <wad@chromium.org>

Review URL: http://codereview.chromium.org/4293001