chromiumos: qcusbnet2k: Take the rcu lock around task list walk.
A more elegant solution to this problem (i.e., having a release hook instead of
a flush hook, and unconditionally freeing in the release hook) does not work. I
am looking into it; in the mean time, this ought to fix the race condition here.

BUG=chromium-os:8882
TEST=Adhoc

Good luck testing this; I've tried stress-testing before and never gotten the
bug to show. There is a stack trace attached to 8882 which this change fixes,
but I can't make it happen again to test with.

Change-Id: Icc4a28701f27b04575603e57f46ccaf6319ef34d
Where: locus chromeos-dogfood no-flush-for-you
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>

Review URL: http://codereview.chromium.org/4750002