Chromium: tpm: Make sure STM TPMs always have self test run on resume
STM TPMs need to have the self test run on resume.  This adds a vendor-specific check to ensure that the test is run.  The check caches the manufacturer ID on the first resume, deferring any penalty of checking this data until then.

To test, on a system with an STM TPM:
1. Make sure the TPM is initialized (`cryptohome --action=tpm_status` will tell you if it is owned, etc.)
2. Log in
3. In another window, `cryptohome --action=status` should show no errors, and the current session TPM wrapped
4. Suspend the system (close the lid)
5. Resume the system (open the lid)
6. Unlock the screen, verify that your session is active
7. In the other window, `cryptohome --action=status`.  With the fix, there should be no errors.  Without the fix, you will see a last error of 0x801, which indicates that self test hasn't been run.

Change-Id: I8638dc6b195a6457eb4fcbe10ed079bfdff747f3
Signed-off-by: Frank Swiderski <fes@chromium.org>

BUG=chromium-os:9553
TEST=see above

Review URL: http://codereview.chromium.org/5373001