CHROMIUM: Call stack trace should use bp when available
In dumpstack_32.c:
When show_registers() calls show_stack_log_lvl, it passes in &regs->sp
for the sp argument but not regs->bp for the bp argument, even though it
is also available.  Instead, it passes in 0.  Later in the call
sequence, when bp=0, it will use the current context's bp.  But this is
not correct during a stack dump based on a provided set of registers,
which are not necessarily from the current context.

The result of this incorrect usage is that one of the stack dumps
becomes completely unreliable due to incompatible bp and sp values.
This patch fixes that by passing regs->bp in show_registers in
dumpstack_32.c when frame pointer is available.  Otherwise the old
behavior is maintained.

As an aside, dumpstack_64.c has always made this call using regs->bp.
It seems like it was developed independently of dumpstack_32.c.

NOTE: this patch is probably not suitable for upstreaming since the
problem will be fixed independently in 2.6.38.  That patch cannot be
easily applied here either due to major restructuring.

Review URL: http://codereview.chromium.org/6001014

Change-Id: Ic517d781c77c0210486847908039e8bab6a8af6d