CHROMIUM: verity: remove verify_context
verify_context is no longer needed now that we guarantee
that all bios are a multiple of PAGE_SIZE in size. Since
there is no longer a ctx.bio we rename the base_bio field
in dm_verity_io to just bio.

The linkages are now as follows: a clone bio has private
pointer to a dm_verity_io which contains a pointer to the
original bio.

BUG=9752
TEST=Booted and visited a few web sites.

Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Change-Id: Ib3e9b56c1b8f1d9f0f8765f40b2b24bd854a4645

Review URL: http://codereview.chromium.org/6033007