CHROMIUM: i915: Backlight to PWM conversions round more accurately
Modified the formulas for backlight to PWM value and vice versa to round
to the nearest integer, rather than rounding down.  The old formula
caused problems with converting from backlight to PWM and then
converting back to backlight.  The result was not always the same as the
original value.

BUG=chromium-os:7615
TEST=Write to i915 backlight brightness and read actual_brightness.
Should match.

Change-Id: Ib38bf3b2184dc88b401992de0d56d07bffc77a3e
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6264006