Revert "UBUNTU: SAUCE: acpi battery -- move first lookup asynchronous"
This reverts commit a969bc97017663585294ec82f36e8a2da7673191
which unfortunately introduces a race condition in
acpi/battery.c that causes a kernel crash.

a969bc was originally made with a comment that said acpi access can take
up to 300ms.  On Cr-48 I measure a delay of around 25ms, deemed "in the noise"
by our boot performance engineer.

a969bc introduced a race condition that initialized part of a battery
data structure, then deferred completion of the initialization until
later.  But an event could come in and complete the init before the
planned async thread, or worse, the event and async thread could try
to complete the init at the same time.

BUG=chromium-os:10475
TEST=Boot system while (un)plugging AC and observe no crash.

Signed-off-by: Bryan Freed <bfreed@chromium.org>

Review URL: http://codereview.chromium.org/6327012

Change-Id: Iaaa52233bea2ad9ae4ac63b34482ef7727d44afb