CHROMIUM: verity: add BUG in check_block for consistency with verify_path
We assume that the caller guarantees that all nodes along the path are
already populated. We already have a BUG_ON checking this in verify_path.
This change adds a similar BUG_ON in check_path and removes the code
which checks the entry state.

BUG=9752
TEST=Ran unit tests in dm-verity.git. Ran platform_DMVerityCorruption on H/W.

Change-Id: Id456d9e495978c93d0be28bd895e5ced28d9973b
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>

Review URL: http://codereview.chromium.org/6246094