CHROMIUM: chromeos_acpi driver reads list of methods from MLST ACPI method
Change the chromeos_acpi driver to read its list of methods from the
MLST ACPI method.  This allows the chromeos_acpi device to support
future expansion of the methods provided by the BIOS, without changing
the device source code each time.

The driver still supports a hard-coded list of methods if the MLST
method is not present; this supports Cr-48 systems, which did not provide MLST.

Signed-off-by: Randall Spangler <rspangler@chromium.org>

BUG=12485
TEST=emerge-x86-generic kernel, then manually test on Cr-48

(1) With existing Cr-48 BIOS, check that /sys/devices/platform/chromeos_acpi
contains the same contents as before the patch.

(2) With a patched BIOS which provides MLST, check that
/sys/devices/platform/chromeos_acpi contains the methods listed in MLST.

Review URL: http://codereview.chromium.org/6576021

Change-Id: I48b9aa9f279fa5cd9acd8ad51629dcd76c334591