CHROMIUM: verity: do all hashing in dm-bht.c
We were doing the hashing of the disk blocks in dm-verity.c and hashing
hash blocks in dm-bht.c. We can simplifiy the code by doing all the
hashing in dm-bht.c. We couldn't do this earlier because we had
to handle unaligned reads (no longer an issue).

I also removed the BUG_ON in dm_bht_get_node because I wanted to use it
in dm_bht_check_block with depth == bht->depth.

BUG=9752
TEST=Ran dm-verity.git unit tests. Ran platform_DMVerityCorruption on H/W.

Note: Had to modify the tests to use the new interface.

Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Change-Id: I246a39b83a727743b49fcb420e57629d69af5dfd

Review URL: http://codereview.chromium.org/6626037