CHROMIUM: verity: remove verify_mode
We don't really use verify_mode any more. We used to use it when
dm_bht_compute called verify_path but that is no longer the case.

I'm sending this patch now because it makes the LOCALLY_VERIFIED state
patch I'm working on a little cleaner. So this is really work
leading up to that.

BUG=9752
TEST=Ran platform_DMVerityCorruption on H/W.

Change-Id: I4ef602f1b0a6c1c4a26927925dbe6d7abe75048c
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>

Review URL: http://codereview.chromium.org/6659037