CHROMIUM: chromeos_acpi Support new NVRAM interface.
This CL changes the interface between chromeos_acpi and
chromeos drivers to accommodate the new NVRAM contents
structure (as described in http://goo.gl/kW1he).

Each shared integer element retrieved through ACPI is placed
in its own container, including the value and a boolean
indicating that the value has been already retrieved. This
scheme obviates the need for a separate
`chromeos_acpi_available` flag.

Writing nvram is now causing recalculating the crc8 value.

We want to be make sure that chromeos.c and chromeos_acpi.c
are always present together. The Kconfig change enforces
this.

The scheme is still vulnerable to race condition when more
than one entity are trying to modify the nvram buffer
contents, the locking mechanism needs to be implemented, but
this is beyond the scope of this change.

Change-Id: I636cc65d6a40ecd2a7b67a4d3bb7d2eb935d8723
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>

BUG=chromium-os:12734
TEST= see below, ideas welcome

The system builds and boots with the appropriate messages
showing up in the kernel log and the NVRAM read failure not
reported:

localhost ~ # dmesg | egrep -i '(chr|Failed reading NVRAM)'
[    0.000000] Linux version 2.6.32.26+drm33.12 (vbendeb@eskimo.mtv.corp.google.com) (gcc version 4.4.3 (gcc-4.4.3_cos_gg_v1_41723) ) #1 SMP Wed Mar 16 18:23:48 PDT 2011 (Ubuntu 2.6.32-0.1-chromeos-intel-menlow)
[    0.106129] checking TSC synchronization [CPU#0 -> CPU#1]: passed.
[    0.178137] checking TSC synchronization [CPU#0 -> CPU#2]:
[    0.403575] chromeos_acpi: installed
[    0.404253] Chrome OS platform detected
[    1.000655] device-mapper: verity-chromeos: dm-verity-chromeos registered

Review URL: http://codereview.chromium.org/6673050