CHROMEOS: chromeos_acpi: handle data buffer overflow.
It was observed that running crossystem on a target with a
recent firmware (supporting the VDAT acpi object) causes
the kernel to complain about buffer overruns (see the
attached bug entry).

It turns out that the recent firmwares supply more than
3700 bytes in VDAT object, which (when converted into an
ASCII string) results in a buffer exceeding maximum sysfs
file size allowance of 4096 (kernel page size). The driver
was truncating the data to exactly the page size, which
caused the kernel generated warnings.

The proper fix would be to split the supplied data into
multiple sysfs objects, each not exceeding 4k, but as of
now only the small part of the VDAT buffer close to its
beginning is used by the application, so it is OK to
truncate the data at source.

This fix moves the VDAT processing code into a separate
function. It now stores the hex dump representation of the
beginning of the buffer guaranteed to fit into 4096 bytes.

Change-Id: I7f380f690be7cf4746f2fa76db37895e58773d73
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>

BUG=chromium-os:13482
TEST=msnual

- build a new image
- install it on the target supporting VDAT in firmware
- restart the target
1. Run the following commands:

dmesg | tail
crossystem
dmesg | tail

observe that crossystem succeeds and there is no messages
added to the dmesg output.

2.
localhost ~ # cat /sys/devices/platform/chromeos_acpi/VDAT | wc
    84    1336    4092
localhost ~ # dmesg | grep chromeos_acpi
[    0.370592] chromeos_acpi: truncating buffer from 3750 to 1336
[    0.371629] chromeos_acpi: installed

Review URL: http://codereview.chromium.org/6732045