CHROMIUM: chromneos_acpi: support both firmware flavors
Recent modification of chromeos_acpi.c (under
http://codereview.chromium.org/6673050) accomodated the new
firmware NVRAM interface, where a multibyte block is
reserved in the NVRAM to communicate between the firmware
and the kernel.

This change is incompatible with the "legacy" ChromeOS
firmware (used in CR-48 devices). Two things got broken for
CR-48:

- it is not considered a ChromeOS platform anymore, because
the availability of the VBNV ACPI element is used to decide
whether the platform is ChromeOS or not, and this element is
not exported by the legacy firmware. if the platform is not
a ChromeOS device, the crash memory dump is not initialized
on start up, hence the bug this CL is associated with.

- even if this was fixed, the
chromeos.c:chromeos_set_need_recovery() interface is also
broken, because it presumes that the nvram buffer is
available.

The fix is to use CHSW ACPI object as the evidence of
running on a ChromeOS platform, and using either CHNV
(if present) or VBNV to communicate the need for recovery
mode reboot.

Change-Id: I5674d68e2eec4e4313783a7edcac1da2b2c9123f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>

BUG=chromium-os:13483
TEST=manual, see below

Testing was conducted on two platforms, Mario (legacy
firmware) and Alex (new firmware).

-1-

. install the new image
. reboot the DUT
. examine the kernel startup log:

on Mario:
localhost ~ # dmesg | egrep 'chromeos_|Chrome'
[    0.470199] chromeos_acpi: falling back to default list of methods
[    0.470249] chromeos_acpi: registering CHSW 0
[    0.471392] chromeos_acpi: registering CHNV 0
[    0.471651] chromeos_acpi: failed to retrieve MLST (5)
[    0.471690] chromeos_acpi: installed
[    0.472069] Legacy ChromeOS firmware detected
localhost ~ #

on Alex:

localhost ~ # dmesg | egrep 'chromeos_|Chrome'
[    0.365498] chromeos_acpi: registering CHSW 0
[    0.366985] chromeos_acpi: registering VBNV 0
[    0.367012] chromeos_acpi: registering VBNV 1
[    0.368926] chromeos_acpi: truncating buffer from 3750 to 1336
[    0.369894] chromeos_acpi: installed
[    0.370490] ChromeOS firmware detected
localhost ~ #

- 2 -

. execute the following commands

localhost ~ # rm -rf /var/spool/crash
localhost ~ # echo -n test-consent > '/home/chronos/Consent To Send Stats'
localhost ~ # reboot

. observe the DUT reboot
. login again and enter

localhost ~ # echo bug > /proc/breakme

. observe the DUT reboot
. after it comes back up

localhost ~ # ls /var/spool/crash
kernel.20110329.180733.0.kcrash  kernel.20110329.180733.0.meta

Review URL: http://codereview.chromium.org/6775012