CHROMIUMOS: usbnet: Kill, then restart interrupt URB on suspend/resume
Otherwise a USB ethernet device that is not re-enumerated on resume
won't resume status polling.  For ASIX this means that device carrier
state isn't monitored and reported after suspend-resume.  A change of
this ilk is already in kernel-next, so I don't believe we need to
upstream this.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:1524
TEST=Reboot with these changes plus code in asix.c to report when the
device status is updated.

Change-Id: I662250a17e62dc5e1dafbc2360010df88fb91472

Review URL: http://codereview.chromium.org/6588096