Enable async suspend/resume of i2c devices
This improves our resume time when we have devices on an i2c bus
that are slow to resume.  In particular we have a light sensor that
adds about 50ms of resume time on one device. We have to enable it
both on the i2c master and i2c client side and then we get fully async
suspend/resume.

BUG=chrome-os-partner:2490
TEST= do a suspend/resume cycle and ensure i2c devices were correctly
resumed

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: I4d8d882aec533d42acf7d0da3a85327ae7ceb97b

Review URL: http://codereview.chromium.org/6689025