ecryptfs: modify write path to encrypt page in writepage
Change the write path to encrypt the data only when the page is written to
disk in ecryptfs_writepage. Previously, ecryptfs encrypts the page in
ecryptfs_write_end which means that if there are multiple write requests to
the same page, ecryptfs ends up re-encrypting that page over and over again.
This patch minimizes the number of encryptions needed.

Change-Id: Ib24b0931005fb577506af5ee11a4ab693a2bfc5f
Signed-off-by: Thieu Le <thieule@chromium.org>
[tyhicks: Changed NULL .drop_inode sop pointer to generic_drop_inode]
Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>

BUG=chromium-os:11706
TEST=fsx-linux, bash-shared-mapping, usemem, git cloned the kernel tree, make defconfig and make, web browsing, change Chrome preferences

Review URL: http://codereview.chromium.org/6823046