CHROMIUM: Input: introduce device properties
Today, userspace sets up an input device based on the data it emits.
This is not always enough; a tablet and a touchscreen may emit exactly
the same data, for instance, but the former should be set up with a
pointer whereas the latter does not need to. Recently, a new type of
touchpad has emerged where the buttons are under the pad, which
changes logic without changing the emitted data. This patch introduces
a new ioctl, EVIOCGPROP, which enables user access to a set of device
properties useful during setup. The properties are given as a bitmap
in the same fashion as the event types, and are also made available
via sysfs, uevent and /proc/bus/input/devices.

Change-Id: Ia55c9f9fc44847090247dcef8cedc6c053a14394
Acked-by: Ping Cheng <pingc@wacom.com>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Dmitry Torokhov <dtor@mail.ru>
Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
(cherry picked from commit aefa83a889256918be11599d82e180ae85457a41)

Signed-off-by: Daniel Stone <daniel.stone@collabora.co.uk>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:13970
TEST=Cr-48 trackpad works as expected with x86-generic build

Review URL: http://codereview.chromium.org/6823007