CHROMIUMOS: Don't return an error for interrupt URB submit
usbnet is currently a mess in terms of suspend-resume.  The resume
function can be called at any time, including during device
discovery of a device that was inserted during system resume.
As architected, it seems impossible to determine whether we
should or should not (re)start interrupt URBs from usbnet_resume
or from usbnet_start.  If we do it from both, the only downside
is that we end up with errors due to the URB already being active.
Unfortunately usb_submit_urb doesn't current return -EINPROGRESS
for this error condition, so for now we need to mute errors from
usb_submit_urb().

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chromium-os:10955
TEST=Run on both a platform that shuts down USB at suspend and one that
does not

Change-Id: I16e6aa6fb71b5f960e34966dc2a3b759bac3ffcc

Review URL: http://codereview.chromium.org/6877020