ath9k_hw: partially revert "fix dma descriptor rx error bit parsing"
The rx error bit parsing was changed to consider PHY errors and various
decryption errors separately. While correct according to the documentation,
this is causing spurious decryption error reports in some situations.

Fix this by restoring the original order of the checks in those places,
where the errors are meant to be mutually exclusive.

If a CRC error is reported, then MIC failure and decryption errors
are irrelevant, and a PHY error is unlikely.

Signed-off-by: Felix Fietkau <nbd@openwrt.org>
Signed-off-by: John W. Linville <linville@tuxdriver.com>

BUG=chromium-os:14507
TEST=SecMat WiFiPerf

Review URL: http://codereview.chromium.org/6883138

Change-Id: Ic5c27ed91d862e8259ba4e5104080e773b9a3806