Disable VMX unless directed not to in the command line
With this patch, the kernel will by default disable VMX on all cores.
This is an improvement over what we had with earlier
systems, which only disabled it only on core 0.

There is a command line parameter called disablevmx.
If it is not specified in the command line VMX will be disabled.
If it is specified with no value, VMX will be disabled.
If it is specified with the value 'on', VMX will be disabled.
If it is specified with a value other than 'off', or 'on', the message
Malformed early option 'disablevmx'
will appear in the log, and VMX will *still* be disabled.

If the value is 'off' VMX will not be disabled.
At the same time, the kernel will
not specifically enable it. It takes no action, leaving
the management decision for later.

This has been tested on a 64-bit platform with and without the option,
with values (both incorrect and correct) for the option, and it works.
I don't have a 32-bit platform to test on.

Without the disablevmx=off option the kernel prints this:
[    0.000000] Disabling VMX on cpu 0
[    0.139521] Disabling VMX on cpu 1
[    0.153338] Disabling VMX on cpu 2
[    0.167148] Disabling VMX on cpu 3

BUG=chrome-os-partner:10449

TEST=build an image for 64-bit Chromeos platform as normal. Boot with as
normal and observe this output:

[    0.000000] Disabling VMX on cpu 0
[    0.139521] Disabling VMX on cpu 1
[    0.153338] Disabling VMX on cpu 2
[    0.167148] Disabling VMX on cpu 3

Add the do_not_disablevmx flag from the command line. There is no output,
but rdmsr commands for register 3a will show 0x0000000000000000, e.g.
localhost ~ # rdmsr 0 0x3a
0x0000000000000000
localhost ~ # rdmsr 1 0x3a
0x0000000000000000
localhost ~ # rdmsr 2 0x3a
0x0000000000000000
localhost ~ # rdmsr 3 0x3a
0x0000000000000000

UNLESS, on, e.g. a lumpy, the lock bit was set, in which case
the msrs will have whatever value they had last boot time.

We had a discussion with a kernel developer about upstreaming this option
and we did not reach closure. I'm assuming we'll be owning this patch
for a while, so I'd like to limit the footprint to one file.

Change-Id: I2aa6f258ffcc04f8d146f878b19f13a805d64f0b
Signed-off-by: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26688
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>

Conflicts:
	arch/x86/kernel/cpu/common.c
[filbranden: 3.18 rebase. Solved trivial conflicts.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>