CHROMIUM: Enable VMX if we are not disabling it.
Previously, we used the disablevmx flag to mean 'disable if not set or
if set to on, do nothing if set to off'. However, since we do not whitelist
the register (0x3c) that would let programs enable vmx, this left us
stranded, with a non-disabled vmx we could not enable. It's simplest
to just enable it and lock it. Only test kernels are built for vmx,
so enabling it is really not going to get anyone very far.

The new meaning:
1. if disablevmx is not set, disable vmx and lock the register
2. if disablevmx is set to on, disable vmx and lock the register
3. if disablevm is set to off, enable vmx and lock the register

I'll test this when it looks like it's in a reasonable form.

BUG=None
TEST=Not tested yet
BRANCH=None

Change-Id: Id3fd5dc401300214e188de48854bd4c6d2777c2b
Signed-off-by: Ronald G. Minnich <rminnich@chromium.org>
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169532
Reviewed-by: Kees Cook <keescook@chromium.org>

Conflicts:
	arch/x86/kernel/cpu/common.c
[filbranden: 3.18 rebase. Solved trivial conflicts.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>